feat: separate ActionDialogManager content
